refactor: centralize slack instructions and share style prefs
