+++ scripts/build_webhook_package.sh
@@ -32,7 +32,6 @@ uv pip install -r requirements-webhook.lock --target "$TEMP_DIR" --no-deps
 echo -e "${YELLOW}Copying webhook source code...${NC}"
 cp -r src/webhook "$TEMP_DIR/"
 cp -r src/shared "$TEMP_DIR/"
-cp -r src/emojismith "$TEMP_DIR/"
 cp src/webhook_handler.py "$TEMP_DIR/"
 cp src/__init__.py "$TEMP_DIR/"
