+++ src/emojismith/domain/value_objects/__init__.py
@@ -1,5 +1,5 @@
 from .emoji_specification import EmojiSpecification
-from .emoji_style_preferences import EmojiStylePreferences
+from shared.domain.value_objects import EmojiStylePreferences
 from .webhook_request import WebhookRequest
 
 __all__ = ["EmojiSpecification", "EmojiStylePreferences", "WebhookRequest"]
